infinite rest fixed

+++ src/main/java/com/system/contact/Repository/PhoneBookRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PhoneBookRepository extends JpaRepository<PhoneBook, Integer> {
+public interface PhoneBookRepository extends JpaRepository<PhoneBook, Long> {
 
     PhoneBook findByName(String name);
     List<PhoneBook> findByDescription(String description);
